cleanup unused files. clippy,
