+++ ol/genesis-tools/src/lib.rs
@@ -8,15 +8,12 @@
     unused_qualifications
 )]
 
-pub mod fetch_archive;
+// pub mod fetch_archive; // Deprecated. Validators will fetch backup zip manually.
 // pub mod fork_daemon; // Note: 0L deleted: This is feature (fork daemon shouldn't be exposed yet, it will not ship with this version of the recovery tool)
 pub mod compare;
 pub mod db_utils;
 pub mod fork_genesis;
 pub mod process_snapshot;
 pub mod read_snapshot;
-pub mod recover;
-pub mod exec_migration;
-// pub mod exec_migration_alt;
-pub mod ol_genesis_context;
+// pub mod recover;
 // pub mod swarm_genesis; // Note: 0L deleted: starting a test environment (Swarm in v5) has changed. Now we would use Forge. And that would be external to main code. Either in ./src/tests or in the Smoke tests directory.
