+++ ol/genesis-tools/tests/export_test.rs
@@ -1,8 +1,9 @@
 mod support;
 
 use ol_genesis_tools::{
-    process_snapshot::db_backup_into_recovery_struct, recover::save_recovery_file,
+    process_snapshot::db_backup_into_recovery_struct,
 };
+use ol_types::legacy_recovery::save_recovery_file;
 use std::fs;
 use support::path_utils::snapshot_path;
 
@@ -20,7 +21,7 @@ async fn test_parse_and_save() {
         .expect("could not export backup into json file");
 
     let output = backup.parent().unwrap().join("test_recovery.json");
-    save_recovery_file(&recovery, &output.to_owned())
+    save_recovery_file(&recovery, &output)
         .expect("ERROR: failed to create recovery from snapshot,");
     fs::remove_file(output).unwrap();
 }
